Move execution plan classes into root of the interpreter module
Implement matching interpreter in submodule and export it explicitly
from interpreter module. No logic change in this commit.
